Make a higher order function for building JSON file instances
